[dependencies] обновил версии зависимостей

+++ src/main/kotlin/ru/kiporskiy/tgbot/librarian/transport/TelegramSender.kt
@@ -64,7 +64,7 @@ object TelegramSender : Sender {
      */
     private fun telegramUserToApplicationUser(tgUser: com.pengrad.telegrambot.model.User): User {
         return User(
-            tgUser.id(),
+            tgUser.id().toInt(),
             tgUser.username() ?: "",
             tgUser.firstName() ?: "",
             tgUser.lastName() ?: ""
